Move serialize tests into submodule
